+++ data/leaf_pairs.csv
@@ -1,5 +1,5 @@
-65,84
-61,183
-75,161
-13,180
-94,159
+96,165
+134,164
+17,90
+161,180
+72,95
